Avoid adding redundant firstnames to user.bin (#21)
Most of the entries in user.bin have a firstname field that matches
the first whitespace-separated part of the name field. Here, we
avoid the need for that by displaying the first part of the name
field if the firstname field is empty.

This reduces the size of user.bin without losing any functionality.