Add tweak: volumebar.enable
If you miss the Tyt volume bargraph display when you turn the Volume
Knob, this tweak is for you.   It may also fix a regression in the
backlight timer, caused by the bargraph patch.  This may or may not be
happening.   This merge adds the missing asterisk in the tweaks menu ,
when the tweak is enabled... Oops.