Fix kd4z.applet.patch to match md380tools changes
Applying root/kd4z.applet.patch failed when attempting to patch
~/md380tools/applet/src/netmon.h.  This is because of changes
in this md380tools commit:
	commit 716da33550ecbf08c63d5ecfacc695c8a5f60b3d
	Author: DL2MF <meinhard.guenther@gmail.com>
	Date:   Sat Jan 6 20:44:01 2018 +0100

	    adding TS to RX display and lastheard

I noticed that the patch contained the entirety of netmon.h, rather
than just the 3 changed lines.  I updated the patch so it contains
the minimal changes to avoid conflicts like this in the future.