Update Canadian province abbreviations
Change "NL" to "Newfoundland and Labrador"
Add "NU" for "Nunavut"

I'd like to add "NT" for "Northwest Territories", but that is already
taken by "Northwest Territory" in Australia.  But there are no
DMR IDs assigned to hams in Northwest Territories, so I'm not going
to fret over it.

Thanks to Tyler Tidman, VA3DGN, for a similar request to
https://github.com/DaleFarnsworth-DMR/userdb